4th Year/[SEG] Proj - Secure, multi-player online domino game in Python/server/requirements.txt
asn1crypto==1.4.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
pycparser==2.20
PyKCS11==1.5.10
pyOpenSSL==20.0.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
